+++ .openzeppelin/project.json
@@ -6,7 +6,9 @@
     "ChargedSettings": "ChargedSettings",
     "ChargedParticles": "ChargedParticles",
     "ChargedManagers": "ChargedManagers",
+    "ParticleSplitter": "ParticleSplitter",
     "GenericWalletManager": "GenericWalletManager",
+    "GenericWalletManagerB": "GenericWalletManagerB",
     "GenericBasketManager": "GenericBasketManager",
     "GenericBasketManagerB": "GenericBasketManagerB",
     "AaveWalltManager": "AaveWalletManager",
