+++ abis/ChargedParticles.json
@@ -586,84 +586,6 @@
   {
     "inputs": [
       {
-        "internalType": "address",
-        "name": "contractAddress",
-        "type": "address"
-      },
-      {
-        "internalType": "uint256",
-        "name": "tokenId",
-        "type": "uint256"
-      },
-      {
-        "internalType": "string",
-        "name": "basketManagerId",
-        "type": "string"
-      },
-      {
-        "internalType": "address",
-        "name": "externalAddress",
-        "type": "address"
-      },
-      {
-        "internalType": "bytes",
-        "name": "encodedParams",
-        "type": "bytes"
-      }
-    ],
-    "name": "executeForBasket",
-    "outputs": [
-      {
-        "internalType": "bytes",
-        "name": "",
-        "type": "bytes"
-      }
-    ],
-    "stateMutability": "payable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
-        "internalType": "address",
-        "name": "contractAddress",
-        "type": "address"
-      },
-      {
-        "internalType": "uint256",
-        "name": "tokenId",
-        "type": "uint256"
-      },
-      {
-        "internalType": "string",
-        "name": "walletManagerId",
-        "type": "string"
-      },
-      {
-        "internalType": "address",
-        "name": "externalAddress",
-        "type": "address"
-      },
-      {
-        "internalType": "bytes",
-        "name": "encodedParams",
-        "type": "bytes"
-      }
-    ],
-    "name": "executeForWallet",
-    "outputs": [
-      {
-        "internalType": "bytes",
-        "name": "",
-        "type": "bytes"
-      }
-    ],
-    "stateMutability": "payable",
-    "type": "function"
-  },
-  {
-    "inputs": [
-      {
         "internalType": "uint256",
         "name": "assetAmount",
         "type": "uint256"
