+++ hardhat.config.js
@@ -152,10 +152,12 @@ module.exports = {
         'ChargedSettings',
         'ChargedManagers',
         'ChargedParticles',
+        'ParticleSplitter',
         'AaveWalletManager',
         'AaveWalletManagerB',
         'AaveBridgeV2',
         'GenericWalletManager',
+        'GenericWalletManagerB',
         'GenericBasketManager',
         'GenericBasketManagerB',
         'Ionx',
